# shared/openapi.yaml
openapi: 3.1.0
info:
  title: Artifact Atlas API
  version: 0.2.0
  description: >
    Local-first asset graph, Artifact BOM, and context-pack builder API.
    All API/storage enum values are canonical and stable; UI labels may differ.
    Sensitive assets default to metadata-only agent access.
    Canonical promotion is human-gated unless policy explicitly permits automation.
  contact:
    name: Artifact Atlas Project
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /api/projects/{projectId}/bom:
    parameters:
      - $ref: "#/components/parameters/projectId"
    get:
      operationId: getProjectBom
      tags: [bom]
      summary: Get the Artifact BOM for a project
      description: Returns the BOM with embedded slots and their assignment counts.
      responses:
        "200":
          description: Project BOM
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bom"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/projects/{projectId}/bom/apply-template:
    parameters:
      - $ref: "#/components/parameters/projectId"
    post:
      operationId: applyBomTemplate
      tags: [bom]
      summary: Apply an artifact template to create or extend the project BOM
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BomApplyTemplateRequest"
      responses:
        "200":
          description: Updated BOM after template application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bom"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/bom/{bomId}:
    parameters:
      - $ref: "#/components/parameters/bomId"
    patch:
      operationId: updateBom
      tags: [bom]
      summary: Update BOM metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BomUpdate"
      responses:
        "200":
          description: Updated BOM
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bom"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/bom/{bomId}/coverage:
    parameters:
      - $ref: "#/components/parameters/bomId"
    get:
      operationId: getBomCoverage
      tags: [coverage]
      summary: Get coverage summary for a BOM
      parameters:
        - name: group_by
          in: query
          schema:
            type: string
            enum: [domain, phase, template]
            default: domain
      responses:
        "200":
          description: Coverage summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoverageSummary"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/bom/{bomId}/gaps:
    parameters:
      - $ref: "#/components/parameters/bomId"
    get:
      operationId: getBomGaps
      tags: [coverage]
      summary: List unfilled or blocked BOM slots
      parameters:
        - name: critical_only
          in: query
          description: Only return required slots that are missing or stale
          schema:
            type: boolean
            default: false
        - name: status
          in: query
          schema:
            type: array
            items:
              $ref: "#/components/schemas/BomSlotStatus"
          style: form
          explode: true
      responses:
        "200":
          description: Gap list
          content:
            application/json:
              schema:
                type: object
                required: [gaps]
                properties:
                  gaps:
                    type: array
                    items:
                      $ref: "#/components/schemas/BomSlot"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/bom/slots/{slotId}/assign:
    parameters:
      - $ref: "#/components/parameters/slotId"
    post:
      operationId: assignSlot
      tags: [bom]
      summary: Assign an asset to a BOM slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SlotAssignRequest"
      responses:
        "201":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BomAssignment"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/bom/slots/{slotId}/mark-not-applicable:
    parameters:
      - $ref: "#/components/parameters/slotId"
    post:
      operationId: markSlotNotApplicable
      tags: [bom]
      summary: Mark a BOM slot as not applicable for this project
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Human-readable justification
      responses:
        "200":
          description: Slot updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BomSlot"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/bom/slots/{slotId}/request-asset:
    parameters:
      - $ref: "#/components/parameters/slotId"
    post:
      operationId: requestAssetForSlot
      tags: [bom]
      summary: Create an asset request/task for an unfilled BOM slot
      description: >
        Emits an atlas_event of type bom_slot_asset_requested and optionally
        creates a linked IntentTree task reference.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
                intenttree_node_id:
                  type: string
                  description: Optional node to link the request to
      responses:
        "202":
          description: Asset request recorded
          content:
            application/json:
              schema:
                type: object
                required: [event_id, slot_id]
                properties:
                  event_id:
                    type: string
                  slot_id:
                    type: string
        "404":
          $ref: "#/components/responses/NotFound"
  /api/bom/assignments/{assignmentId}:
    parameters:
      - name: assignmentId
        in: path
        required: true
        schema:
          type: string
    delete:
      operationId: unassignSlot
      tags: [bom]
      summary: Remove an asset assignment from a BOM slot
      description: >
        Tombstones the assignment. If no accepted assignments remain on the slot,
        the slot status reverts to partial or missing. Emits a bom_slot_filled
        audit event.
      parameters:
        - name: reason
          in: query
          schema:
            type: string
          description: Optional justification for removal
      responses:
        "204":
          description: Assignment removed
        "404":
          $ref: "#/components/responses/NotFound"
  /api/bom/assignments/{assignmentId}/status:
    parameters:
      - name: assignmentId
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: updateAssignmentStatus
      tags: [bom]
      summary: Update the assignment_status of an existing slot assignment
      description: >
        Promotes a suggested assignment to accepted or canonical, or rejects it.
        Advancing from suggested to accepted/canonical also advances the slot
        status from partial to in_progress. Emits a bom_slot_filled audit event.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assignment_status]
              properties:
                assignment_status:
                  $ref: "#/components/schemas/AssignmentStatus"
      responses:
        "200":
          description: Updated assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BomAssignment"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

  # ---------------------------------------------------------------------------
  # Context Packs
  # ---------------------------------------------------------------------------
...
components:
...
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      description: Unique project identifier
      schema:
        type: string
        example: proj_agentic_os
...
    bomId:
      name: bomId
      in: path
      required: true
      description: Unique BOM identifier
      schema:
        type: string
        example: bom_proj_agentic_os
    slotId:
      name: slotId
      in: path
      required: true
      description: Unique BOM slot identifier
      schema:
        type: string
        example: slot_architecture_diagram
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: not_found
              message: "The requested resource was not found."
              request_id: "req_abc123"
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorEnvelope"
...
  schemas:
...
    BomSlotStatus:
      type: string
      enum: [missing, partial, in_progress, complete, stale, blocked, not_applicable]
      description: Canonical BOM slot fill status
...
    AssignmentStatus:
      type: string
      enum: [suggested, accepted, rejected, canonical]
      description: Canonical BOM slot assignment status
...
    BomMergeStrategy:
      type: string
      enum: [new_section, merge_existing, phase_only, node_subtree]
    SlotPhase:
      type: string
      enum: [discovery, design, build, launch, operate, review, custom]
      nullable: true
...
    AssignedBy:
      type: string
      enum: [user, agent, system]
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              examples:
                - not_found
                - policy_denied
                - conflict
                - internal_error
            message:
              type: string
              description: Human-readable error description
            detail:
              type: object
              description: Additional structured context (varies by error type)
              nullable: true
            request_id:
              type: string
              nullable: true
    ValidationErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, validation_errors]
          properties:
            code:
              type: string
              enum: [validation_error]
            message:
              type: string
            request_id:
              type: string
              nullable: true
            validation_errors:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field:
                    type: string
                    description: Dot-separated field path
                  message:
                    type: string
                  value:
                    description: The invalid value (if safe to include)

    # ---- Projects ----
...
    Bom:
      type: object
      required: [id, project_id, name, status, merge_strategy, coverage_score, created_at, updated_at]
      properties:
        id:
          type: string
          example: bom_proj_agentic_os
        project_id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [active, archived]
        source_templates:
          type: array
          items:
            type: string
          description: Template IDs that were applied to compose this BOM
        merge_strategy:
          $ref: "#/components/schemas/BomMergeStrategy"
        coverage_score:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: Fraction of required slots with at least one accepted/canonical assignment
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        slots:
          type: array
          items:
            $ref: "#/components/schemas/BomSlot"
          description: Embedded slots (may be empty when called without expand parameter)
    BomUpdate:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, archived]
    BomApplyTemplateRequest:
      type: object
      required: [template_id]
      properties:
        template_id:
          type: string
        merge_strategy:
          $ref: "#/components/schemas/BomMergeStrategy"
          default: merge_existing
        intenttree_node_id:
          type: string
          nullable: true
          description: Scope template application to a subtree
    BomSlot:
      type: object
      required: [id, bom_id, artifact_type_id, domain, required, status, min_assets]
      properties:
        id:
          type: string
          example: slot_architecture_diagram
        bom_id:
          type: string
        artifact_type_id:
          type: string
        domain:
          type: string
          example: Architecture
        phase:
          $ref: "#/components/schemas/SlotPhase"
        required:
          type: boolean
        status:
          $ref: "#/components/schemas/BomSlotStatus"
        linked_intenttree_node_id:
          type: string
          nullable: true
        min_assets:
          type: integer
          minimum: 0
          default: 1
        max_assets:
          type: integer
          nullable: true
        staleness_days:
          type: integer
          nullable: true
        guidance:
          type: string
          nullable: true
        assignment_count:
          type: integer
          nullable: true
          description: Total assignments (all statuses)
        accepted_assignment_count:
          type: integer
          nullable: true
          description: Assignments with status accepted or canonical
    BomAssignment:
      type: object
      required: [id, slot_id, asset_id, assignment_status, assigned_by, assigned_at]
      properties:
        id:
          type: string
        slot_id:
          type: string
        asset_id:
          type: string
        assignment_status:
          $ref: "#/components/schemas/AssignmentStatus"
        confidence:
          type: number
          format: float
          nullable: true
          minimum: 0.0
          maximum: 1.0
        assigned_by:
          $ref: "#/components/schemas/AssignedBy"
        assigned_at:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
    SlotAssignRequest:
      type: object
      required: [asset_id]
      properties:
        asset_id:
          type: string
        slot_id:
          type: string
          description: Required when calling from /api/assets/{assetId}/assign-slot
        assignment_status:
          $ref: "#/components/schemas/AssignmentStatus"
          default: suggested
        confidence:
          type: number
          format: float
          nullable: true
        notes:
          type: string
          nullable: true

    # ---- Coverage ----
    CoverageSummary:
      type: object
      required: [bom_id, coverage_score, total_slots, filled_slots, missing_slots, stale_slots]
      properties:
        bom_id:
          type: string
        coverage_score:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
        total_slots:
          type: integer
        required_slots:
          type: integer
        filled_slots:
          type: integer
        missing_slots:
          type: integer
        stale_slots:
          type: integer
        blocked_slots:
          type: integer
        not_applicable_slots:
          type: integer
        groups:
          type: array
          description: Coverage broken down by group_by parameter (domain, phase, or template)
          items:
            type: object
            required: [group_key, coverage_score, total_slots, filled_slots, missing_slots]
            properties:
              group_key:
                type: string
                description: Domain name, phase, or template ID depending on group_by
              coverage_score:
                type: number
                format: float
              total_slots:
                type: integer
              filled_slots:
                type: integer
              missing_slots:
                type: integer

    # ---- Context Packs ----
...